♻️ refactor : CouponWallet 에서 CouponRedemption 네이밍 리팩토링
