+++ src/main/java/com/coumin/woowahancoupons/coupon/controller/CouponRedemptionRestController.java
@@ -3,6 +3,6 @@ package com.coumin.woowahancoupons.coupon.controller;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class CouponWalletRestController {
+public class CouponRedemptionRestController {
 
 }
