+++ src/main/java/com/coumin/woowahancoupons/coupon/service/CouponRedemptionService.java
@@ -2,6 +2,6 @@ package com.coumin.woowahancoupons.coupon.service;
 
 import java.util.UUID;
 
-public interface CouponWalletService {
+public interface CouponRedemptionService {
     void allocateCouponToCustomer(UUID couponId, Long customerId);
 }
