+++ src/main/java/com/coumin/woowahancoupons/domain/coupon/CouponRedemptionRepository.java
@@ -3,6 +3,6 @@ package com.coumin.woowahancoupons.domain.coupon;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CouponWalletRepository extends JpaRepository<CouponWallet, UUID> {
+public interface CouponRedemptionRepository extends JpaRepository<CouponRedemption, UUID> {
 
 }
